pyproject.toml
[tool.poetry]
name = "databricks-ddbxutils"
version = "0.1.0"
description = "extends databricks dbutils"
authors = ["Haneul Kim <haneul.kim@data-dynamics.io>"]
readme = "README.md"
packages = [{ include = "ddbxutils" }]

[tool.poetry.dependencies]
python = "^3.11"
databricks-sdk = "^0.57.0"
dotenv = "^0.9.9"
jinja2 = "^3.1.6"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
